+++ test/rave.charge.test.js
@@ -231,4 +231,511 @@ describe('#Rave charge', function () {
     expect(resp.data).to.have.property('auth_model', 'TOKEN');
     expect(resp.meta.authorization).to.have.property('mode', 'redirect');
   });
+
+  it('should return Unauthorized Googlepay charge', async function () {
+    this.timeout(10000);
+
+    const createGooglePayCharge = sinon.stub(chargeInstance, 'googlepay').resolves({
+      "status": "error",
+      "message": "Merchant is not enabled to accept googlepay transactions.",
+      "data": null
+    })
+
+    var payload = {
+      "tx_ref": "MC-TEST-1234568_success_mock",
+      "amount": "10",
+      "currency": "USD",
+      "email": "user@example.com",
+      "fullname": "Yolande Aglaé Colbert",
+      "redirect_url": "https://flutterwave.ng",
+      "client_ip": "192.168.0.1",
+      "device_fingerprint": "gdgdhdh738bhshsjs",
+      "billing_zip": "15101",
+      "billing_city": "allison park",
+      "billing_address": "3563 Huntertown Rd",
+      "billing_state": "Pennsylvania",
+      "billing_country": "US",
+      "meta": {
+          "metaname": "testmeta",
+          "metavalue": "testvalue"
+      }
+    };
+    var resp = await chargeInstance.googlepay(payload);
+    expect(createGooglePayCharge).to.have.been.calledOnce;
+    expect(createGooglePayCharge).to.have.been.calledOnceWith(payload);
+
+    expect(resp).to.have.property('status', 'error');
+    expect(resp).to.have.property('message', 'Merchant is not enabled to accept googlepay transactions.');
+    expect(resp).to.have.property('data', null);
+  });
+
+  it('should return Googlepay charge', async function () {
+    this.timeout(10000);
+
+    const createGooglePayCharge = sinon.stub(chargeInstance, 'googlepay').resolves({
+      "status": "success",
+      "message": "Charge initiated",
+      "data": {
+          "id": 2615403,
+          "tx_ref": "MC-TEST-1234568_success_mock",
+          "flw_ref": "RQFA6549001367743",
+          "device_fingerprint": "gdgdhdh738bhshsjs",
+          "amount": 10,
+          "charged_amount": 10,
+          "app_fee": 0.38,
+          "merchant_fee": 0,
+          "processor_response": "Payment token retrieval has been initiated",
+          "auth_model": "GOOGLEPAY_NOAUTH",
+          "currency": "USD",
+          "ip": "54.75.56.55",
+          "narration": "Test Google Pay charge",
+          "status": "pending",
+          "auth_url": "https://rave-api-v2.herokuapp.com/flwv3-pug/getpaid/api/short-url/XPtNw-WkQ",
+          "payment_type": "googlepay",
+          "fraud_status": "ok",
+          "charge_type": "normal",
+          "created_at": "2022-05-11T20:36:15.000Z",
+          "account_id": 20937,
+          "customer": {
+              "id": 955307,
+              "phone_number": null,
+              "name": "Yolande Aglaé Colbert",
+              "email": "user@example.com",
+              "created_at": "2022-05-11T20:36:14.000Z"
+          },
+          "meta": {
+              "authorization": {
+                  "mode": "redirect",
+                  "redirect": "https://rave-api-v2.herokuapp.com/flwv3-pug/getpaid/api/short-url/XPtNw-WkQ"
+              }
+          }
+      }
+  });
+
+    var payload = {
+      "tx_ref": "MC-TEST-1234568_success_mock",
+      "amount": "10",
+      "currency": "USD",
+      "email": "user@example.com",
+      "fullname": "Yolande Aglaé Colbert",
+      "redirect_url": "https://flutterwave.ng",
+      "client_ip": "192.168.0.1",
+      "device_fingerprint": "gdgdhdh738bhshsjs",
+      "billing_zip": "15101",
+      "billing_city": "allison park",
+      "billing_address": "3563 Huntertown Rd",
+      "billing_state": "Pennsylvania",
+      "billing_country": "US",
+      "meta": {
+          "metaname": "testmeta",
+          "metavalue": "testvalue"
+      }
+    };
+    var resp = await chargeInstance.googlepay(payload);
+    expect(createGooglePayCharge).to.have.been.calledOnce;
+    expect(createGooglePayCharge).to.have.been.calledOnceWith(payload);
+
+    expect(resp).to.have.property('status', 'success');
+    expect(resp).to.have.property('message', 'Charge initiated');
+
+    expect(resp.data).to.have.property('auth_model', 'GOOGLEPAY_NOAUTH');
+    expect(resp.data.meta.authorization).to.have.property('mode', 'redirect');
+  });
+
+  it('should return ValidationError for Googlepay charge', async function () {
+    this.timeout(10000);
+
+    var payload = {
+      "tx_ref": "MC-TEST-1234568_success_mock",
+      "amount": "10",
+      "email": "user@example.com",
+      "fullname": "Yolande Aglaé Colbert",
+      "redirect_url": "https://flutterwave.ng",
+      "client_ip": "192.168.0.1",
+      "device_fingerprint": "gdgdhdh738bhshsjs",
+      "billing_zip": "15101",
+      "billing_city": "allison park",
+      "billing_address": "3563 Huntertown Rd",
+      "billing_state": "Pennsylvania",
+      "billing_country": "US",
+      "meta": {
+          "metaname": "testmeta",
+          "metavalue": "testvalue"
+      }
+    };
+
+    try{
+    var resp = await chargeInstance.googlepay(payload);
+    }catch(err){
+      expect(err).to.be.returned;
+      expect(err.message).to.include('"currency" is required');
+    }
+  });
+
+  it('should return Unauthorized applepay charge', async function () {
+    this.timeout(10000);
+
+    const createApplePayCharge = sinon.stub(chargeInstance, 'applepay').resolves({
+      "status": "error",
+      "message": "Merchant is not enabled for ApplePay collections.",
+      "data": null
+    })
+
+    var payload = {
+      "tx_ref":"MC-TEST-123456",
+      "amount":"10",
+      "currency":"USD",
+      "email": "user@example.com",
+      "fullname": "Yolande Aglaé Colbert",
+      "redirect_url":"https://flutterwave.ng",
+      "client_ip":"192.168.0.1",
+      "device_fingerprint":"gdgdhdh738bhshsjs",
+      "billing_zip":"15101",
+      "billing_city":"allison park",
+      "billing_address":"3563 Huntertown Rd",
+      "billing_state":"Pennsylvania",
+      "billing_country":"US",
+      "phone_number":"09012345678",
+      "meta":{
+          "metaname":"testmeta",
+          "metavalue":"testvalue"
+      }
+    };
+    var resp = await chargeInstance.applepay(payload);
+    expect(createApplePayCharge).to.have.been.calledOnce;
+    expect(createApplePayCharge).to.have.been.calledOnceWith(payload);
+
+    expect(resp).to.have.property('status', 'error');
+    expect(resp).to.have.property('message', 'Merchant is not enabled for ApplePay collections.');
+    expect(resp).to.have.property('data', null);
+  });
+
+  it('should return Applepay charge', async function () {
+    this.timeout(10000);
+
+    const createApplePayCharge = sinon.stub(chargeInstance, 'applepay').resolves({
+      "status":"success",
+      "message":"Charge initiated",
+      "data":{
+         "id":645498756,
+         "tx_ref":"MC-TEST-1234523",
+         "flw_ref":"TKVH48681032738026",
+         "device_fingerprint":"gdgdhdh738bhshsjs",
+         "amount":1,
+         "charged_amount":1.04,
+         "app_fee":0.04,
+         "merchant_fee":0,
+         "processor_response":"Pending validation",
+         "auth_model":"APPLEPAY",
+         "currency":"GBP",
+         "ip":"192.168.0.1",
+         "narration":"Test payment",
+         "status":"pending",
+         "auth_url":"https://applepay.aq2-flutterwave.com?reference=TKVH48681032738026",
+         "payment_type":"applepay",
+         "fraud_status":"ok",
+         "charge_type":"normal",
+         "created_at":"2022-06-11T12:18:11.000Z",
+         "account_id":3442,
+         "customer":{
+            "id":379560157,
+            "phone_number":"09012345678",
+            "name":"Flutterwave Developers",
+            "email":"developers@flutterwavego.com",
+            "created_at":"2022-06-11T12:18:11.000Z"
+         },
+         "meta":{
+            "authorization":{
+               "mode":"redirect",
+               "redirect":"https://applepay.aq2-flutterwave.com?reference=TKVH48681032738026"
+            }
+         }
+      }
+   })
+
+    var payload = {
+      "tx_ref":"MC-TEST-123456",
+      "amount":"10",
+      "currency":"USD",
+      "email": "user@example.com",
+      "fullname": "Yolande Aglaé Colbert",
+      "redirect_url":"https://flutterwave.ng",
+      "client_ip":"192.168.0.1",
+      "device_fingerprint":"gdgdhdh738bhshsjs",
+      "billing_zip":"15101",
+      "billing_city":"allison park",
+      "billing_address":"3563 Huntertown Rd",
+      "billing_state":"Pennsylvania",
+      "billing_country":"US",
+      "phone_number":"09012345678",
+      "meta":{
+          "metaname":"testmeta",
+          "metavalue":"testvalue"
+      }
+    };
+    var resp = await chargeInstance.applepay(payload);
+    expect(createApplePayCharge).to.have.been.calledOnce;
+    expect(createApplePayCharge).to.have.been.calledOnceWith(payload);
+
+    expect(resp).to.have.property('status', 'success');
+    expect(resp).to.have.property('message', 'Charge initiated');
+
+    expect(resp.data).to.have.property('auth_model', 'APPLEPAY');
+    expect(resp.data.meta.authorization).to.have.property('mode', 'redirect');
+  });
+
+  it('should return ValidationError for Applepay charge', async function () {
+    this.timeout(10000);
+
+    var payload = {
+      "tx_ref":"MC-TEST-123456",
+      "amount":"10",
+      "email": "user@example.com",
+      "fullname": "Yolande Aglaé Colbert",
+      "redirect_url":"https://flutterwave.ng",
+      "client_ip":"192.168.0.1",
+      "device_fingerprint":"gdgdhdh738bhshsjs",
+      "billing_zip":"15101",
+      "billing_city":"allison park",
+      "billing_address":"3563 Huntertown Rd",
+      "billing_state":"Pennsylvania",
+      "billing_country":"US",
+      "phone_number":"09012345678",
+      "meta":{
+          "metaname":"testmeta",
+          "metavalue":"testvalue"
+      }
+    };
+
+    try{
+    var resp = await chargeInstance.applepay(payload);
+    }catch(err){
+      expect(err).to.be.returned;
+      expect(err.message).to.include('"currency" is required');
+    }
+  });
+
+  it('should return Unauthorized eNaira charge', async function () {
+    this.timeout(10000);
+
+    const createeNairaCharge = sinon.stub(chargeInstance, 'enaira').resolves({
+      "status": "error",
+      "message": "Merchant is not enabled for eNaira collections.",
+      "data": null
+    })
+
+    var payload = {
+      "tx_ref":"MC-TEST-123456",
+      "amount":"100",
+      "currency":"NGN",
+      "email":"user@example.com",
+      "fullname":"Yemi Desola",
+      "phone_number":"09000000000",
+      "redirect_url":"https://flutterwave.ng"
+    };
+    var resp = await chargeInstance.enaira(payload);
+    expect(createeNairaCharge).to.have.been.calledOnce;
+    expect(createeNairaCharge).to.have.been.calledOnceWith(payload);
+
+    expect(resp).to.have.property('status', 'error');
+    expect(resp).to.have.property('message', 'Merchant is not enabled for eNaira collections.');
+    expect(resp).to.have.property('data', null);
+  });
+
+  it('should return eNaira charge', async function () {
+    this.timeout(10000);
+
+    const createeNairaCharge = sinon.stub(chargeInstance, 'enaira').resolves({
+      "status": "success",
+      "message": "Charge initiated",
+      "data": {
+          "id": 4197118,
+          "tx_ref": "12345test_05",
+          "flw_ref": "ZZYO0021678723801871881",
+          "device_fingerprint": "N/A",
+          "amount": 200,
+          "charged_amount": 200,
+          "app_fee": 2.8,
+          "merchant_fee": 0,
+          "processor_response": "pending",
+          "auth_model": "ENAIRA",
+          "currency": "NGN",
+          "ip": "54.75.161.64",
+          "narration": "Flutterwave Developers",
+          "status": "pending",
+          "payment_type": "enaira",
+          "fraud_status": "ok",
+          "charge_type": "normal",
+          "created_at": "2023-03-13T16:10:00.000Z",
+          "account_id": 20937,
+          "customer": {
+              "id": 1953337,
+              "phone_number": "08092269174",
+              "name": "Wisdom Joshua",
+              "email": "wsdmjsh@gmail.com",
+              "created_at": "2023-01-18T13:22:14.000Z"
+          },
+          "meta": {
+              "authorization": {
+                  "mode": "redirect",
+                  "redirect": "https://camltest.azurewebsites.net/enairapay/?invoiceId=01GVDVRTG80MVSRJJQQYRFTZK3&amount=200&token=438890"
+              }
+          }
+      }
+  })
+
+    var payload = {
+      "tx_ref":"MC-TEST-123456",
+      "amount":"100",
+      "currency":"NGN",
+      "email":"user@example.com",
+      "fullname":"Yemi Desola",
+      "phone_number":"09000000000",
+      "redirect_url":"https://flutterwave.ng"
+    };
+    var resp = await chargeInstance.enaira(payload);
+    expect(createeNairaCharge).to.have.been.calledOnce;
+    expect(createeNairaCharge).to.have.been.calledOnceWith(payload);
+
+    expect(resp).to.have.property('status', 'success');
+    expect(resp).to.have.property('message', 'Charge initiated');
+
+    expect(resp.data).to.have.property('auth_model', 'ENAIRA');
+    expect(resp.data.meta.authorization).to.have.property('mode', 'redirect');
+  });
+
+  it('should return ValidationError for eNaira charge', async function () {
+    this.timeout(10000);
+
+    var payload = {
+      "tx_ref":"MC-TEST-123456",
+      "amount":"100",
+      "email":"user@example.com",
+      "fullname":"Yemi Desola",
+      "phone_number":"09000000000",
+      "redirect_url":"https://flutterwave.ng"
+    };
+
+    try{
+    var resp = await chargeInstance.enaira(payload);
+    }catch(err){
+      expect(err).to.be.returned;
+      expect(err.message).to.include('"currency" is required');
+    }
+  });
+
+  it('should return Fawrypay charge', async function () {
+    this.timeout(10000);
+
+    const createFawryPayCharge = sinon.stub(chargeInstance, 'fawrypay').resolves({
+      "status": 'success',
+      "message": 'Charge initiated',
+      "data": {
+        "id": 4511926,
+        "tx_ref": 'fawrySample1',
+        "order_ref": 'URF_FAWRY_1691396932978_2086335',
+        "flw_ref": '9263673349',
+        "device_fingerprint": 'N/A',
+        "amount": 10,
+        "charged_amount": 10,
+        "app_fee": 0.23,
+        "merchant_fee": 0,
+        "processor_response": 'Request is pending',
+        "currency": 'EGP',
+        "narration": 'Flutterwave Developers',
+        "status": 'pending',
+        "auth_url": 'N/A',
+        "payment_type": 'fawry_pay',
+        "fraud_status": 'ok',
+        "charge_type": 'normal',
+        "created_at": '2023-08-07T08:28:52.000Z',
+        "account_id": 20937,
+        "customer": {
+          "id": 1869436,
+          "phone_number": '09012345678',
+          "name": 'Anonymous customer',
+          "email": 'user@flw.com',
+          "created_at": '2022-10-30T22:25:31.000Z'
+        }
+      },
+      "meta": {
+        "authorization": {
+          "mode": 'fawry_pay',
+          "instruction": 'Please make payment with the flw_ref returned in the response which should be the same as the reference sent via SMS'
+        }
+      }
+    })
+
+    var payload = {
+      "tx_ref": "fawrySample1",
+      "amount": "10",
+      "email": "user@flw.com",
+      "currency": "EGP",
+      "phone_number": "09012345678",
+      "redirect_url": "https://www.flutterwave.com",
+      "meta": {
+          "name": "Cornelius"
+      }
+    };
+    var resp = await chargeInstance.fawrypay(payload);
+    expect(createFawryPayCharge).to.have.been.calledOnce;
+    expect(createFawryPayCharge).to.have.been.calledOnceWith(payload);
+
+    expect(resp).to.have.property('status', 'success');
+    expect(resp).to.have.property('message', 'Charge initiated');
+
+    expect(resp.data).to.have.property('payment_type', 'fawry_pay');
+    expect(resp.meta.authorization).to.have.property('mode', 'fawry_pay');
+  });
+
+  it('should return Unauthorized Fawrypay charge', async function () {
+    this.timeout(10000);
+
+    const createFawryPayCharge = sinon.stub(chargeInstance, 'fawrypay').resolves({
+      "status": "error",
+      "message": "Merchant is not enabled for Fawry Pay collections.",
+      "data": null
+    })
+
+    var payload = {
+      "tx_ref": "fawrySample1",
+      "amount": "10",
+      "email": "user@flw.com",
+      "currency": "EGP",
+      "phone_number": "09012345678",
+      "redirect_url": "https://www.flutterwave.com",
+      "meta": {
+          "name": "Cornelius"
+      }
+    };
+    var resp = await chargeInstance.fawrypay(payload);
+    expect(createFawryPayCharge).to.have.been.calledOnce;
+    expect(createFawryPayCharge).to.have.been.calledOnceWith(payload);
+
+    expect(resp).to.have.property('status', 'error');
+    expect(resp).to.have.property('message', 'Merchant is not enabled for Fawry Pay collections.');
+    expect(resp).to.have.property('data', null);
+  });
+
+  it('should return ValidationError for Fawrypay charge', async function () {
+    this.timeout(10000);
+
+    var payload = {
+      "tx_ref": "fawrySample1",
+      "amount": "10",
+      "email": "user@flw.com",
+      "phone_number": "09012345678",
+      "redirect_url": "https://www.flutterwave.com",
+      "meta": {
+          "name": "Cornelius"
+      }
+    };
+
+    try{
+    var resp = await chargeInstance.fawrypay(payload);
+    }catch(err){
+      expect(err).to.be.returned;
+      expect(err.message).to.include('"currency" is required');
+    }
+  });
 });